requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.6
jieba3k==0.35.1
lxml==4.1.1
newspaper3k==0.2.6
nltk==3.2.5
Pillow==5.0.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.22
